consolidate viz marker tests
